+++ libft/ft_memccpy.c
@@ -6,11 +6,11 @@
 /*   By: aymaatou <aymaatou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/14 16:42:18 by aymaatou          #+#    #+#             */
-/*   Updated: 2019/10/14 16:55:33 by aymaatou         ###   ########.fr       */
+/*   Updated: 2019/10/16 13:15:03 by aymaatou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memccpy(void *dest, const void *src, int c, unsigned int n)
+void	*ft_memccpy(void *dest, const void *src, int c, size_t n)
 {
 	unsigned int i;
